requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
httpx>=0.25.0
sse-starlette>=1.8.0
pydantic>=2.5.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
Pillow>=10.0.0

# Authentication
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0
email-validator>=2.0.0

# WebSocket support
websockets>=12.0

# File processing
pypdf>=4.0.0

# Knowledge base (vector search)
numpy>=1.24.0

# Video generation (Hugging Face Spaces)
gradio-client>=0.10.0

# Image generation (browser automation for HuggingFace Spaces)
playwright>=1.40.0

# Caching with TTL and size limits
cachetools>=5.3.0
aiohttp>=3.9.0
